requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-monitor-query>=1.3.0

# Optional: Azure OpenAI for intelligent error analysis
# openai>=1.12.0
